agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "ipykernel (>=6.29.5,<7.0.0)",
    "jupyter (>=1.1.1,<2.0.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
    "psycopg2 (>=2.9.10,<3.0.0)",
    "pgvector (>=0.4.1,<0.5.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "fastapi (>=0.115.14,<0.116.0)",
    "google-api-python-client (>=2.174.0,<3.0.0)",
    "google-generativeai (>=0.8.5,<0.9.0)",
    "requests (>=2.32.4,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
jupyter = "^1.1.1"

